# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Eliassen.AspNetCore.Mvc",
    "version": "1.0.0.0"
  },
  "paths": {
...
    "/api/Documents/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-permissions": {
          "anonymous": false
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-System.Boolean"
                }
              }
            }
          }
        },
        "x-permissions": {
          "anonymous": false,
          "right": [
            "right_lesson_admin",
            "right_blog_admin",
            "right_project_admin"
          ]
        }
      }
    },
    "/api/Documents": {
      "post": {
        "tags": [
          "Documents"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object"
              },
              "encoding": {
                "DocumentId": {
                  "style": "form"
                },
                "DocumentKey": {
                  "style": "form"
                },
                "DocumentName": {
                  "style": "form"
                },
                "DocumentType": {
                  "style": "form"
                },
                "DocumentSize": {
                  "style": "form"
                },
                "DocumentRepository": {
                  "style": "form"
                },
                "DocumentCategory": {
                  "style": "form"
                },
                "CreatedOn": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-Nucleus.Core.Contracts.Models.DocumentModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-Nucleus.Core.Contracts.Models.DocumentModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.ResponseModel-Nucleus.Core.Contracts.Models.DocumentModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        },
        "x-permissions": {
          "anonymous": false,
          "right": [
            "right_lesson_admin",
            "right_blog_admin",
            "right_project_admin"
          ]
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "Nucleus.Core.Contracts.Models.DocumentModel": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "nullable": true
          },
          "documentKey": {
            "type": "string",
            "nullable": true
          },
          "documentName": {
            "type": "string",
            "nullable": true
          },
          "documentType": {
            "type": "string",
            "nullable": true
          },
          "documentSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "documentRepository": {
            "type": "string",
            "nullable": true
          },
          "documentCategory": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Nucleus.Core.Contracts.Models.ResponseModel-Nucleus.Core.Contracts.Models.DocumentModel": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "$ref": "#/components/schemas/Nucleus.Core.Contracts.Models.DocumentModel"
          }
        },
        "additionalProperties": false
      },
...
      "Nucleus.Core.Contracts.Models.ResponseModel-System.Boolean": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "B2C authentication",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://lightwellnucleusdev.b2clogin.com/lightwellnucleusdev.onmicrosoft.com/oauth2/v2.0/authorize?p=B2C_1_NucAcc_Signin",
            "tokenUrl": "https://lightwellnucleusdev.b2clogin.com/lightwellnucleusdev.onmicrosoft.com/oauth2/v2.0/token?p=B2C_1_NucAcc_Signin",
            "scopes": {
              "https://lightwellnucleusdev.onmicrosoft.com/user.read/read": "UserReadApiClaim"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [ ]
    }
  ]
}